SaveDump with bad characters
stackoverflow.com/questions/20310904/filename-chars-cause-opening-failed-einval-invalid-argument

Please refer to this page, you will find the full list.

http://en.wikipedia.org/wiki/Comparison_of_file_systems

Although that your sdcard might be internal and mount as FUSE, it might be still FAT32 filesystem. The valid characters for FAT32 are (according to the page):

Any byte except for values 0-31, 127 (DEL) and: " * / : < > ? \ | + , . ; = [] (lowcase a-z are stored as A-Z). With VFAT LFN any Unicode except NUL[14][15